+++ playground/nextjs-app-router/components/demo/TransactionDefault.tsx
@@ -2,7 +2,8 @@ import { useCapabilities } from '@/lib/hooks';
 import { clickCalls, clickContracts } from '@/lib/transactions';
 import { TransactionDefault } from '@coinbase/onchainkit/transaction';
 import { useCallback, useContext, useEffect } from 'react';
-import { AppContext, TransactionTypes } from '../AppProvider';
+import { AppContext } from '../AppProvider';
+import { TransactionTypes } from '@/types/onchainkit';
 
 export default function TransactionDefaultDemo() {
   const { chainId, transactionType } = useContext(AppContext);
